Move SourceType enum to Journal.
